fix RPGDemo

+++ RPGDemo/Controllers/EnemyState/EnemyAttackState.cs
@@ -14,7 +14,7 @@ public class EnemyAttackState : IState<Controller>
         controller.Character.ComboNumber = 0;
         Vector2 joyDir = Vector2.Zero;
         controller.Character.Move(ref joyDir);
-        controller.Character.SetAnimation((int)Character.AnimationIndices.Attack);
+        controller.Character.SetAnimation(Character.AnimationIndices.Attack);
     }
 
     public virtual void Exit(Controller controller)
